Create SpendingStats component
